test: make the adoption bench a faithful 2-mode GUI proxy (#325)
Collapse the bench from 5 modes to exactly two — speckit (stock upstream)
and companion — matching the post-#312 workflow consolidation. MODES /
PRESET_BY_MODE / PROFILE_BY_MODE / armNote / writeVscodeSettings / report
rendering all drop the turbo/lean/fast-path/logs rungs; legacy rows in
stats.jsonl / history.jsonl still read without crashing, they just stop
being generated.

Faithful dispatch: extract the pure preamble renderers out of
promptBuilder.ts into a new vscode-free module promptPreamble.ts
(renderPreamble / renderLifecyclePreamble /
renderSpecifyCreationLifecyclePreamble + the schema/status/timing blocks).
promptBuilder.ts imports and re-exports them, keeping the vscode config
toggle and nowUtc() — its public API is unchanged. The bench imports the
compiled dist/ai-providers/promptPreamble.js via a new driver.mjs so it
prepends the EXACT same per-step preamble the GUI dispatches, for BOTH
modes, and can never drift from the real path. check-shape-parity stays
green (byte-identical output).

Settle-wait: add waitForSettle(cellDir, step, timeoutMs) in lib.mjs that
polls the cell's .spec-context.json for the step's completed-form status
before advancing (the same signal the GUI watchers wait on), with a unit
test (waitForSettle.test.mjs, node --test, no AI).

Isolate capture overhead: measureCell reads captureOverheadSec from the
cell's .run-meta.json and surfaces it as its own row in the stats row and
the rendered report, separate from work time.

README reframed near the top: agent runs are far faster than human GUI
runs, so absolute wall-clock won't match; the bench delivers a trustworthy
RELATIVE stock-vs-companion comparison plus isolated capture overhead, and
the user's own GUI runs remain the absolute yardstick.

Verified: npm run compile + npm test (1000) green; waitForSettle unit test
green; run-all.mjs --dry-run shows 2 modes; check-shape-parity OK.
