+++ docs/capture-and-timing.md
@@ -64,7 +64,8 @@ The 2026-06-08 fix above made per-task capture *reliable* (the hook owns it) but
 - **Backstop hardened.** `sync_tasks()` writes finish-only too and now journals per-task **even when implement already self-closed** (`status: implemented`): its guard was narrowed from `TERMINAL_STATUSES` to `CROSS_STEP_TERMINAL` (`completed`/`archived` only), so a same-step write is never rejected. `_journaled_tasks` keeps it from duplicating a task the live path captured.
 - **Parallel `[P]` caveat.** The delta model can't give each task in a parallel batch its own duration — the batch is attributed to whichever finishes last. Accepted for the common sequential case.
 - **Derivation.** `src/features/specs/stepHistoryDerivation.ts` (`buildSubsteps`) computes each row from finish deltas, staying tolerant of legacy start+complete pairs and single boundary markers. A per-task row is named from `task` (or a legacy `substep` mirror); the derivation compares `task` so two distinct finishes don't collapse, and excludes task entries when deciding whether the step itself is closed.
-- **Both dispatch surfaces.** The instruction lives identically in `presets/_parts/timing.md` (spec-kit path, assembled into the bodies by `build-commands.py`) and `src/ai-providers/promptBuilder.ts` (GUI path); `check-shape-parity.py` guards against a fork. The relocation from `_shared/timing-partial.md` to `_parts/timing.md` changed only where the block is authored — the bytes the AI receives, and therefore the capture, are unchanged.
+- **Both dispatch surfaces.** The instruction lives identically in `presets/_parts/timing.md` (spec-kit path, assembled into the bodies by `build-commands.py`) and the GUI preamble renderers in `src/ai-providers/promptPreamble.ts` (GUI path); `check-shape-parity.py` guards against a fork. The relocation from `_shared/timing-partial.md` to `_parts/timing.md` changed only where the block is authored — the bytes the AI receives, and therefore the capture, are unchanged.
+- **GUI preamble is vscode-free + reusable.** The pure preamble renderers (`renderPreamble`, `renderLifecyclePreamble`, `renderSpecifyCreationLifecyclePreamble`, plus the schema/status/timing blocks) live in `src/ai-providers/promptPreamble.ts` with **no `import vscode`** — the only vscode-dependent parts (the `aiContextInstructions` on/off toggle and `nowUtc()`) stay in `promptBuilder.ts`, which imports and re-exports the renderers so its public API is unchanged. This lets the adoption bench import the compiled `dist/ai-providers/promptPreamble.js` and dispatch each step with the *exact* GUI preamble (`bench/driver.mjs`), so the bench can't drift from the real dispatch path.
 - **Both marker formats.** Per-task detection accepts the Companion **bold** form (`- [x] **T001**`) *and* the stock tasks-template **plain** form (`- [x] T001 …`) — `parse_task_markers` (`write-context.py`) and the eval make the `**` optional. Previously the bold-only regex silently no-op'd on a stock `tasks.md`, so the spec got no per-task journal and implement never auto-closed. The TS/GUI task-percent already counted plain checkboxes, so only the Python parsers needed it.
 
 ## Record-shape hardening (2026-06-09)
